requirements.txt
langchain-openai
langchain-core
python-dotenv
streamlit

